[TASK] refactored api actions
