+++ Web/typo3conf/ext/sessions/Classes/Controller/SessionModuleController.php
@@ -178,8 +178,7 @@ class SessionModuleController extends ActionController
 
     public function indexAction()
     {
-        $this->view->assign('linkSessions', $this->listSessionsAction());
-        $this->view->assign('linkRooms', $this->listRoomsAction());
+
     }
 
     /**
@@ -223,104 +222,6 @@ class SessionModuleController extends ActionController
         return $sessions;
     }
 
-    /**
-     * Get all sessions for FullCalendar
-     *
-     * @return String
-     */
-    public function listSessionsAction()
-    {
-        $allSessionFlatArray = $this->sessionRepository->findAllFlat();
-
-        $result = array();
-
-        // Session properties
-        $sessionWhitelist = array('uid', 'title', 'date', 'begin', 'end', 'room', 'description');
-        // Generate session data
-        foreach($allSessionFlatArray as $session)
-        {
-            $tempSessionArray = array();
-            foreach($sessionWhitelist as $property)
-            {
-                $tempSessionArray[$property] = $session[$property];
-            }
-//            $result[] = $tempSessionArray;
-            $result[] = array('id' => $session['uid'],
-                    'resourceId' => $session['room'],
-                    'start' => $session['begin'],
-                    'end' => $session ['end'],
-                    'title' => $session ['title'],
-                    'description' => $session ['description']);
-        }
-        return json_encode($result);
-
-    }
-
-
-    /**
-     * Get all rooms for FullCalendar
-     *
-     * @return String
-     */
-    public function listRoomsAction()
-    {
-        $allRoomFlatArray = $this->roomRepository->findAllFlat();
-
-        $result = array();
-
-        // Generate room data
-        foreach($allRoomFlatArray as $room)
-        {
-            $result[] = array('id' => $room['uid'], 'title' => $room['title']);
-        }
-        return json_encode($result);
-
-    }
-
-    public function initializeUpdateSessionAction()
-    {
-        $this->adjustSessionPropertyMappingConfiguration($this->resourceArgumentName);
-
-        if ($this->request->hasArgument('secondSession')) {
-            $this->adjustSessionPropertyMappingConfiguration('secondSession');
-        }
-    }
-
-    /*public function errorAction(){
-        var_dump($this->request->getOriginalRequestMappingResults()->forProperty('session')->getFlattenedErrors());
-    }*/
-
-    /**
-     * Update session
-     *
-     * @param \TYPO3\Sessions\Domain\Model\ScheduledSession $session
-     * @param \TYPO3\Sessions\Domain\Model\ScheduledSession $secondSession
-     * @return string
-     */
-    public function updateSessionAction($session, $secondSession = null)
-    {
-       /* // Find session object
-        $sessionObject = $this->sessionRepository->findByUid($session['uid']);
-        // Set session properties
-
-        if(!is_null($sessionObject)) {
-            foreach ($session as $propertyName => $value) {
-                if($propertyName != 'uid')
-                    $sessionObject->{'set'.ucwords($propertyName)}($value);
-            }
-            $this->sessionRepository->update($sessionObject);
-        }
-        $this->redirect('listSessionAndRoomsForCalender');*/
-
-        //var_dump($session);die;
-        $this->sessionRepository->update($session);
-        if($secondSession != null) {
-            $this->sessionRepository->update($secondSession);
-        }
-//        return $session->getTitle();
-        return 'success';
-    }
-
     public function generateFirstScheduleAction()
     {
 
@@ -354,21 +255,6 @@ class SessionModuleController extends ActionController
 	    $this->redirect('index', '', '', array('incompleteSessions' => $incompleteSessions, 'creationDone' => true));
     }
 
-    protected function adjustSessionPropertyMappingConfiguration($propertyName) {
-        /** @var \TYPO3\CMS\Extbase\Mvc\Controller\MvcPropertyMappingConfiguration $propertyMappingConfiguration */
-        $propertyMappingConfiguration = $this->arguments[$propertyName]->getPropertyMappingConfiguration();
-
-        $propertyMappingConfiguration->setTypeConverterOption(PersistentObjectConverter::class,
-            PersistentObjectConverter::CONFIGURATION_MODIFICATION_ALLOWED, true);
-        $propertyMappingConfiguration->forProperty('begin')->setTypeConverterOption(DateTimeConverter::class,
-            DateTimeConverter::CONFIGURATION_DATE_FORMAT, DATE_ISO8601);
-        $propertyMappingConfiguration->forProperty('end')->setTypeConverterOption(DateTimeConverter::class,
-            DateTimeConverter::CONFIGURATION_DATE_FORMAT, DATE_ISO8601);
-        $propertyMappingConfiguration->allowProperties('date', 'begin', 'end', 'room', 'title');
-        $propertyMappingConfiguration->skipUnknownProperties();
-
-    }
-
     /*public function errorAction(){
         var_dump($this->request->getOriginalRequestMappingResults()->forProperty('session')->getFlattenedErrors());
     }*/
@@ -376,12 +262,6 @@ class SessionModuleController extends ActionController
     public function testAction()
     {
 
-    }
-
-    public function demoAction()
-    {
-        $session = $this->sessionRepository->findByUid(1);
-        $this->view->assign('session', $session);
     }
 
     /**
