+++ Web/typo3conf/ext/sessions/Resources/Private/Templates/SessionModule/Index.html
@@ -1,3 +1,3 @@
 <h1>Session Module</h1>
 
-<div id="calendar" data-update-session="{f:uri.action(action:'updateSession')}" data-list-sessions="{f:uri.action(action:'listSessions')}" data-list-rooms="{f:uri.action(action:'listRooms')}"></div>
+<div id="calendar" data-update-session="{f:uri.action(action:'updateSession', controller:'ApiModule')}" data-list-sessions="{f:uri.action(action:'listSessions', controller:'ApiModule')}" data-list-rooms="{f:uri.action(action:'listRooms', controller:'ApiModule')}"></div>
